Updated tests.
